+++ backend/tests/integration/tests/permissions/test_user_role_permissions.py
@@ -4,7 +4,7 @@ This file tests the ability of different user types to set the role of other use
 from danswer.db.models import UserRole
 from tests.integration.common_utils.user import TestUser
 from tests.integration.common_utils.user import UserManager
-from tests.integration.common_utils.user_groups import UserGroupManager
+from tests.integration.common_utils.user_group import UserGroupManager
 
 
 def test_user_role_setting_permissions(reset: None) -> None:
