+++ backend/tests/integration/tests/permissions/test_whole_curator_flow.py
@@ -8,7 +8,7 @@ from tests.integration.common_utils.connector import ConnectorManager
 from tests.integration.common_utils.credential import CredentialManager
 from tests.integration.common_utils.user import TestUser
 from tests.integration.common_utils.user import UserManager
-from tests.integration.common_utils.user_groups import UserGroupManager
+from tests.integration.common_utils.user_group import UserGroupManager
 
 
 def test_whole_curator_flow(reset: None) -> None:
